Add `geof:area` and `geof:metricArea` using `GeometryInfo` (#2430)
Add support for the GeoSPARQL functions `geof:area` and `geof:metricArea`. For a `POLYGON` this is computed via `S2Polygon::GetArea()`. For a `LINESTRING` or `POINT`, this is always zero. For more complex shapes it is the sum of the areas of the simpler shapes forming the complex shape. For example (area of the water reservoir "Llac d'Engolasters" in Andorra):
```sparql
SELECT * WHERE {
  osmway:6593464 geo:hasGeometry/geo:asWKT ?geometry .
  BIND (geof:area(?geometry, unit:M2) AS ?area_sqm)
  BIND (geof:area(?geometry, unit:KiloM2) AS ?area_sqkm)
  BIND (geof:area(?geometry, unit:MI2) AS ?area_sqmi)
  BIND (geof:metricArea(?geometry) AS ?metric_area)
}
```
NOTE: This is an index-breaking change with regard to the `on-disk-compressed-geo-split` vocabulary type. It is merged in a sequence with #2238 and #2172