Fix runtime information and live update for lazy join of two index scans (#2480)
Since #2256, the details of the runtime information of the index scans of a lazy join of two index scans were not properly updated anymore. In particular, `num-blocks-read` and `num-elements-read` were always zero, even after completion of the query. A live update of the runtime information of index scans in this scenario never worked so far. This is all fixed now. Together with #2473, live updates now fully work for any cascade of lazy joins. Here is simple example query, where we can now see live updates for all operations (the predicate must be large enough so that the joins are actually lazy):
```sparql
SELECT * {
  ?s a ?o1, ?o2, ?o3
}
```

Co-authored-by: RobinTF <83676088+RobinTF@users.noreply.github.com>