Fix two nasty bugs in `CompressedExternalIdTable` (#2250)
We stumbled upon this when building an index for a dataset, where all subjects are blank nodes. Then the assertion `(numBlocksPushed_ == 0) != compressAndWriteFuture_.valid()` failed. In the process of fixing the bug, another bug showed its face. Both bugs are now fixed as follows:

1. We make sure that `compressAndWriteFuture_` is always set and valid, also for empty blocks for which there is nothing to do. This fixes the assertion error
2. The `ql:has-pattern` relation is read twice during index building and the respective generators must not be alive at the same time, which, however, could happen so far. This is now fixed by ensuring that the first generator is destroyed before the second one is created

While at it, add more checks to ensure that a proper error message is issued in case more unexpected behavior occurs.

Co-authored-by: Johannes Kalmbach <johannes.kalmbach@gmail.com>