Include time for parsing in `JSON` returned for update operations (#2463)
Parsing an update request can take significant time, in particular, for `INSERT DATA` and `DELETE DATE` requests with many triples. So far, the parsing time was not reported in the JSON returned upon completion of an update operation. Now it is. 

Technically, `TimeTracer` is now passed as a `shared_ptr` instead of as `TimeTracer&` on the "outer levels" of the call stack because there were issues with `TimeTracer&`. On the side, fix a bug in `TimeTracer::reset`.