Allow custom `Permutation` and `LocatedTriplesSnapshot` in `IndexScan` (#2519)
Previously, the `IndexScan` constructor took a `Permutation::Enum` and retrieved the respective `Permutation` and `LocatedTriplesSnapshot` from its `QueryExecutionContext`. However, this was limited to the six standard permutations PSO, POS, SPO, SOP, OSP, and OPS. With this change, the `IndexScan` constructor takes a `PermutationPtr` and `LocatedTriplesSnapshotPtr`, which are shared pointers to the "real thing". This can then be used also for custom permutations, which are not part of the regular index, in particular those from the materialized views in #2482

NOTE: The "normal" and "internal" variant of the PSO or POS permutation still use the same `PermutationPtr`. The distinction is made via `std::unique_ptr<Permutation> internalPermutation_` in the `Permutation` class, which is rather hacky. Instead, the distinction should be made in the query planner, so that the `PermutationPtr` object is already the correct permutation (normal or internal). That is work for a separate PR.