Make labeled blank nodes from different files distinct (#2554)
When the same blank node label (e.g., `_:b1`) appears in different RDF files, they must be treated as distinct blank nodes, which so far, QLever did not do. There was already a distinct blank node prefix for each file, but it was used only for anonymous blank nodes. Now it is also used for labeled blank nodes. Fixes #642