Only buffer single payload for WebSocket communication (#2548)
So far, each payload generated by a single query (the JSON strings with the current runtime information) was stored in the `QueryToSocketDistributor` class and sent one after the other to the consumer. Now only the latest payload is stored and when the other side is slow to consume, intermediate payloads are not sent. This saves network traffic and RAM (though the latter is only significant for complex and long-running queries, which generate many large JSON strings in the course of their processing)