Move `ValuesForTesting` back to `test` (#2568)
In the recent #2565, `ValuesForTesting` was moved to `src` because despite its name it wasn't only used for testing. It turned out that the only use outside of testing was for the implementation of `Describe`. This is now implemented using (the anyway more appropriate) `ExplicitIdTableOperation`, and `ValuesForTesting` is back to where it belongs