Implement a serializer, which serializes the data in compressed blocks (#2526)
Implement a generic `CompressedWriteSerializer`, which serializes the data in compressed block and the corresponding `CompressedReadSerializer` for reading such data. Implement a concrete pair `ZstdWriteSerializer` and `ZstdReadSerializer`, which uses these with the `Zstd` compression library.  This will be used to read and write a `NamedCachedResult` such that it can be precomputed on one machine, and used on another

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>