Refactor `CompressedRelationWriter::createPermutationPair` (#2566)
Refactor the (rather lengthy and complex) `CompressedRelationWriter::createPermutationPair` method into an own `struct`, where complex lambdas become proper member functions and the core loop that iterates over all blocks of triples is now much easier to read and much shorter. This is the foundation of further refactoring, in particular for the writing of only a single permutation as opposed to a pair of permutations. Preparation for #2537 and thus also #2482