Optionally add query time to `sparql-results+json` result (#2594)
There is now a runtime parameter `sparql-results-json-with-time`, which when `true` (which is the default), adds a top-level field `meta` with subfields `query-time-ms` and `result-size-total` to the `sparql-results+json` result. The fields `head` and `results` are just like before. Note that the SPARQL 1.1 standard allows the addition of other fields as long as they don't interfere with the mandatory fields. When the runtime parameter is set, the `send` parameter limits the number of results that are sent, just like for `qlever-results+json`, but `"result-size-total"` still provides the total number of results. For example:
```json
{
  "head": { ... }
  "results": { ... }
  "meta": {
    "query-time-ms": 59,
    "result-size-total": 17623
  }
}
```