Allow writing a single permutation with `CompressedRelationWriter` (#2537)
This continues the refactoring from #2566 (which provides a much more modular version of the previously very complex `CompressedRelationWriter::createPermutationPair`), by now providing two methods `createPermutationPair` (which creates a permutation pair, like before) and `createPermutation` (which creates a single permutation). Both methods call the same underlying `writePermutation`, which contains the actual implementation and is now a helper function.

Being able to write a single permutation is useful for a number of further developments, in particular, #2408 (binary index rebuild) and #2482 (materialized views)