Add code for creating, querying, and managing materialized views (#2601)
In particular, implement the new classes `MaterializedView` (a single materialized view, which can be used for index scans), `MaterializedViewWriter` (create a new materialized view from an already planned query, and write it to disk), and `MaterializedViewsManager` (part of a `QueryExecutionContext`, manages all currently loaded materialized views).

This continues #2581 and is preparation for #2482, which will integrate the new functionality into `libqlever` and the `Server` class.

NOTE: We should change the prefix `https://qlever.cs.uni-freiburg.de/` to `https://qlever.dev/` (for all magic services) in a separate PR, either following this one, or directly following #2482