Introduce new `RuntimeInformation::Status::fullyMaterializedCompleted` (#2608)
Rename three of the existing states and add one new. Here are the new names, the last one being for the new state. The new state helps the new Qlue UI to figure out when a lazily materialized operation is completed (which was hard so far when it was the root operation of a query).
```
RuntimeInformation::Status::fullyMaterializedInProgress
RuntimeInformation::Status::fullyMaterializedCompleted
RuntimeInformation::Status::lazilyMaterializedInProgress
RuntimeInformation::Status::lazilyMaterializedCompleted
```