Enable serialization for certain vocabularies (#2586)
Implement serialization functions for the different vocabulary implementations. In particular, `VocabularyInMemory` can be serialized now, as well as `CompressedVocabulary`, `UnicodeVocabulary`, and `PolymorphicVocabulary` when the underlying vocabulary can be serialized. For `VocabularyOnDisk`, or other vocabularies using that as underlying vocabulary, an exception will be thrown because that vocabulary is already on disk