Optimize `GraphTerm::toTripleComponent` (#2603)
So far, the `TripleComponent::parseTripleObject` method created a dummy triple, which it parsed and then extracted the object from. Now simply parse the object (it seems the method for that wasn't available when the previous code was written). In `GraphTerm::toSparql` use this method for parsing literals (which require non-trivial escaping), but use a simpler method for parsing IRIs. Before this change, the expensive parsing method was used for both literals and IRIs